+++ src/Hacks/Level/Noclip/Noclip.hpp
@@ -3,6 +3,7 @@
 #include "../../../Client/Module.hpp"
 #include "../../../Client/FloatSliderModule.hpp"
 #include "../../../Client/ColourModule.hpp"
+#include "../../../Client/InputModule.hpp"
 
 using namespace geode::prelude;
 
@@ -58,7 +59,147 @@ class NoclipTintColour : public ColourModule
         }
 };
 
+
+
+class NoclipPlayer1 : public Module
+{
+    public:
+        MODULE_SETUP(NoclipPlayer1)
+        {
+            setID("noclip/player1");
+            setPriority(4);
+            setDefaultEnabled(true);
+        }
+};
+
+class NoclipPlayer2 : public Module
+{
+    public:
+        MODULE_SETUP(NoclipPlayer2)
+        {
+            setID("noclip/player2");
+            setPriority(5);
+            setDefaultEnabled(true);
+        }
+};
+
+class NoclipPlayer1UseMinAcc : public Module
+{
+    public:
+        MODULE_SETUP(NoclipPlayer1UseMinAcc)
+        {
+            setID("noclip/player1/use-min-accuracy");
+            setPriority(1);
+        }
+};
+
+class NoclipPlayer2UseMinAcc : public Module
+{
+    public:
+        MODULE_SETUP(NoclipPlayer2UseMinAcc)
+        {
+            setID("noclip/player2/use-min-accuracy");
+            setPriority(1);
+        }
+};
+
+class NoclipPlayer1MinAcc : public InputModule
+{
+    public:
+        MODULE_SETUP(NoclipPlayer1MinAcc)
+        {
+            setID("noclip/player1/min-accuracy");
+            setPriority(2);
+
+            setDefaultString("100");
+            setPlaceholderString("Percent");
+            setHint("percentage");
+
+            setStringFilter("1234567890.");
+            setMaxCharCount(5);
+        }
+};
+
+class NoclipPlayer2MinAcc : public InputModule
+{
+    public:
+        MODULE_SETUP(NoclipPlayer2MinAcc)
+        {
+            setID("noclip/player2/min-accuracy");
+            setPriority(2);
+
+            setDefaultString("100");
+            setPlaceholderString("Percent");
+            setHint("percentage");
+
+            setStringFilter("1234567890.");
+            setMaxCharCount(5);
+        }
+};
+
+class NoclipPlayer1UseMaxDeaths : public Module
+{
+    public:
+        MODULE_SETUP(NoclipPlayer1UseMaxDeaths)
+        {
+            setID("noclip/player1/use-max-deaths");
+            setPriority(3);
+        }
+};
+
+class NoclipPlayer2UseMaxDeaths : public Module
+{
+    public:
+        MODULE_SETUP(NoclipPlayer2UseMaxDeaths)
+        {
+            setID("noclip/player2/use-max-deaths");
+            setPriority(3);
+        }
+};
+
+class NoclipPlayer1MaxDeaths : public InputModule
+{
+    public:
+        MODULE_SETUP(NoclipPlayer1MaxDeaths)
+        {
+            setID("noclip/player1/max-deaths");
+            setPriority(4);
+
+            setDefaultString("0");
+            setPlaceholderString("Deaths");
+
+            setStringFilter("1234567890");
+            setMaxCharCount(5);
+        }
+};
+
+class NoclipPlayer2MaxDeaths : public InputModule
+{
+    public:
+        MODULE_SETUP(NoclipPlayer2MaxDeaths)
+        {
+            setID("noclip/player2/max-deaths");
+            setPriority(4);
+
+            setDefaultString("0");
+            setPlaceholderString("Deaths");
+
+            setStringFilter("1234567890");
+            setMaxCharCount(5);
+        }
+};
+
 SUBMIT_HACK(Noclip);
 SUBMIT_OPTION(Noclip, NoclipTintOnDeath);
 SUBMIT_OPTION(Noclip, NoclipTintOpacity);
 SUBMIT_OPTION(Noclip, NoclipTintColour);
+SUBMIT_OPTION(Noclip, NoclipPlayer1);
+SUBMIT_OPTION(Noclip, NoclipPlayer2);
+SUBMIT_OPTION(NoclipPlayer1, NoclipPlayer1UseMinAcc);
+SUBMIT_OPTION(NoclipPlayer2, NoclipPlayer2UseMinAcc);
+SUBMIT_OPTION(NoclipPlayer1, NoclipPlayer1MinAcc);
+SUBMIT_OPTION(NoclipPlayer2, NoclipPlayer2MinAcc);
+SUBMIT_OPTION(NoclipPlayer1, NoclipPlayer1UseMaxDeaths);
+SUBMIT_OPTION(NoclipPlayer2, NoclipPlayer2UseMaxDeaths);
+SUBMIT_OPTION(NoclipPlayer1, NoclipPlayer1MaxDeaths);
+SUBMIT_OPTION(NoclipPlayer2, NoclipPlayer2MaxDeaths);